qdoc: Introduce 'none' selector for .qhp subprojects
By default, all pages in the documentation project are included in a
.qhp subproject if it doesn't define a selector property.

Introduce a new selector, `none`, that matches with nothing. This is
useful for creating subprojects that contain no pages under it, but
simply links to the page defined in the `indexTitle` property.

Fixes: QTBUG-132533
Change-Id: Iff35621ae18d20ff7e8727494695ffbe6be57236
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 6a654c883c71b22cefcfdd650a8dd3d3d1542402)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>