linguist doc: add missing links to id based translation docs
The linguist manual for ID based documentation briefly describes the
ID based translation functions. Some of the functions link to their
reference documentations. But for some reason, several functions are
explicitly annotated by "\c" keyword to format the text in the
monospace font and avoid linking. This is the case
since at least 2012 (the oldest available commit on this content) when
the document file was moved from qtdoc to qttools.
This patch removes the "\c" annotation to re-establish the references
to the function documentations.

Fixes: QTBUG-132707
Change-Id: I861917ac8eb131bcd565158fe27f23c5991b26a1
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
(cherry picked from commit 90af4075080f59e35c0a7b89519ded602d11d55f)