linguist: adjust colors when working with multiple languages
When using multiple languages in linguist, for each language a color
is specified, and the items of that language are presented with the
color specific to that language.
For this purpose, linguist has a set of hardcoded colors that are
defined based on the assumption that the user is using light mode.
As a result, when using dark mode and having multiple languages,
the items for some languages are not readable, and linguist looks
weird.

Adjust the predefined set of colors to the used theme (light or
dark mode). Also react to the theme changes during runtime and
re-adjust the color set.

Fixes: QTBUG-54203
Fixes: QTBUG-136578
Change-Id: Iabb045b0e52440603592ef67b9fb73cd3e8e0d67
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 46259522181446863bbf15abe3c2825b03aa58e9)