CMake: Improve logic to discover suitable LLVM/Clang packages for QDoc
LLVM has a strict compatibility check in the CMake packages they
provide. Therefore, we cannot rely on passing
`QDOC_MINIMUM_CLANG_VERSION` as the VERSION argument to the
`find_package()` call for Clang or LLVM.

To circumvent this, iterate over the versions that are known to be
supported by QDoc. The list of supported versions is maintained next to
the minimum version requirement, such that they can hopefully be kept in
sync in the future.

Pick-to: 6.8
Change-Id: I205128d22ae68e6fed8e108d7c7753d7f59626b7
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit 914e00bbd6b35a376b6f71c031d9469bd4d9f672)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>