lupdate: remove clang parser
Clang is a full-blown C++ parser, which means that lupdate/clang
cannot skip over uninteresting things. It's needs the full
picture of the project, including include paths and defines.
While, lupdate only needs to understand the context of
a translatable string, and nothing more.

As a result, not only the clang-based parser in lupdate has
slower performance compared to the default parser, also,
it must retrieve the compiler's default include
paths and platform defines, and all this is sometimes breaking
unexpectedly (the autotests for lupdate/clang have often been
the culprit of flakiness in the qttools repository).

Also, lupdate/clang can only see the code that is
compiled for the current platform, whereas the default parser
extracts all of these strings:

 \#if defined(Q_OS_WIN)
     show(tr("I'm on Windows!"));
 \#elif defined(Q_OS_LINUX)
     show(tr("I'm on Linux!"));
 \#else
     show(tr("I'm on some other OS!"));
 \#endif

Altogether, because of the points above, lupdate/clang has
no users, and makes maintenance more difficult. Removing
it simplifies the tool and avoids these problems.

Fixes: QTBUG-130096
Fixes: QTBUG-131481
Fixes: QTBUG-130006
Change-Id: I6b41f59f97af58cdb793ec797c8718f78e8bc8a0
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>