qdoc: Improve handling and presentation of anonymous enums
QDoc has a rudimentary support for documenting anonymous C++
enums by recognizing `anonymous` enum name as a special case:

  \enum QFoo::anonymous

Caveat: with documentation in implementation files, only a
single anonymous \enum can be referred to in a specific
namespace. If there are multiple ones, enumerators from each
can be documented with \value commands in the same \enum topic.

When documentation comments are in headers, QDoc failed to
auto-tie a documentation comment preceding an anonymous enum
to its declaration. Fix this by generating a unique name for
the enum based on its location in the header, and adding
a new `isAnonymous` field to EnumNode class.

Based on this new field, drop the auto-generated unique
name of an enum from generated documentation, in favor of
`anonymous` extra synopsis tag.

In a summary section, because there is no name listed which
would normally act as a link to details, use the type name
`enum` as the hyperlink.

In DocBook generator, drop the <db:enumname> element for
anonymous enums.

Add a test for anonymous enums for both scenarios;
documentation in implementation files, and in headers.

Add a test for Utilities::uniqueIdentifier().

[ChangeLog][QDoc] QDoc now supports documenting anonymous C++
enums when `documentationinheaders` feature is enabled.

Fixes: QTBUG-133371
Change-Id: I93bdb31a2de042d980ebd07ec87cc74e9dfb63aa
Reviewed-by: Paul Wicking <paul.wicking@qt.io>