linguist: fix redundant restart prompt in "Search and Translate"
When using the Search and Translate dialog (from the Edit menu),
if the search reaches the last item in the list, Linguist prompts
the user to restart the search from the beginning. After approval,
Linguist may randomly display the same prompt again, despite
the last item not being reached a second time.
Ensures that the restart prompt is only shown when the search
actually wraps around.

Fixes: QTBUG-133320
Change-Id: I5d71c659f4d42190cde1e4aad7034ccb4e1f75ef
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>