qdoc: Write proxy nodes to the index after their members
When QDoc cannot locate an aggregate (namespace/class/header) passed as
an argument to the \relates command, it creates a ProxyNode instance to
serve as an intermediate parent for the related API entity, with the
assumption that the real documented aggregate is located in some other
module. These 'dangling' related non-members are then associated with
the real aggregate when its documentation is generated (related-by-
proxy).

Because of the way these relations are recorded into the index file, it
is important that the member nodes (e.g. global functions) are
constructed first, before constructing the ProxyNode aggregate that then
adopts them.

Postpone writing proxies in QDocIndexFiles::generateIndexSections() to
ensure correct order when reading the index back in, similar to what's
being done for group and module nodes.

Pick-to: 6.8 6.5
Change-Id: I690e543afe949551e6643aba9d7a78b976abc088
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 1d2e5515afd8b76c9263e737f33e4451f3d00c81)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>