allow mixing id-based and text-based translation in a document
Previously, lrelease required the -idbased flag to determine whether
a TS file used ID-based or text-based translation. This imposed a
limitation, as a document had to be entirely ID-based or entirely
text-based.

This patch removes the need for manually setting or unsetting the
-idbased flag. Instead, lrelease now examines each message individually
and determines whether it is ID-based (if an ID is set) or
text-based (if not).

Also, the qsTr in qml (or tr in cpp) functions used to assume text
based translation, even if id is using "//=" notations set. This leads
to wrong translations if the same ids are used in different contexts.
This patch switches the message to id based if id is set.

[ChangeLog][lrelease] The -idbased flag is not required anymore, and
will be removed in future version.

Change-Id: I23cf7b0fdae52b93ac15c552e1bd41e569ac3cb4
Fixes: QTBUG-52180
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>