Revert "Revert "qdoc: Relocate warning for invalid QML property types""
This reverts commit 1152c7a9a96cc5068735559ba7d31bfc14de19ae.

Reason for revert: Reverted one of two patches, this one being the follow-up. The other change remains merged. As we're now addressing the issues that cause the warning, we won't revert the change that introduced this warning. However, this change that was reverted is required for QDoc not to issue duplicates of said warning.

Change-Id: I659f360c8b28644c8adbdef12c0c7ec7dd3abe62
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>