lupdate: unify parsing of meta strings in different languages
Qt linguist allows meta strings in form of comments in the source
code to provide more info for translation. Sofar lupdate used
different meta string parsers for different supported languages
(python, qml, c++, java). This leaded to outdated meta string
parsers for some languages and meta strings.
As a result, there was lack of support for some meta strings
in Qml, Java, and Python.
To bring consistency in parsing meta strings, this patch
provides a parser for meta strings which is used for all
supported languages.

Also in the course of changes in this path, an existing bug in
the python parser is resolved, where the parser recorded
occasionally wrong line numbers for translations.

Change-Id: Iad43fdf905e6b9dcaefe71b71139073d6daf1827
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>