support label attribute in UI files
provide support for label attribute in ui files. The label
is used to categorize ID-based translations on Linguist.
Upon setting a label in the form header, all the id based translations
in that form will acquire the given label in the form.
The same changes are also prepared to be pushed in
qtbase (change number 628224)

Change-Id: I8ef4b06f2633977910b2cabe83bb37137fb6f84c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>