QDoc: Turn "Unknown base for QML type" warning into report
When processing the documentation for QML types, QDoc attempts to
resolve their base type when applicable. In some cases, the base type
may be a dependency. However, if so, and the dependency isn't met (as is
the case in Qt's CI system which integrates repositories separately),
QDoc currently issues a false positive warning. As a consequence, Qt's
documentation configuration contains a spurious regexp that lets QDoc
omit/ignore this warning.

By changing the warning to a report, we ensure that QDoc will notify the
user of the possible mistake. However, this message category doesn't
have an impact on QDoc's warning count, and as such will not contribute
to failing documentation builds in Qt's CI. However, it will be visible
in the build logs instead of being silently ignored. While a modest
improvement, it's still an improvement to the current situation. Another
benefit is that the message will be issued once for the base type,
instead of for each unresolved case.

Pick-to: 6.8
Task-number: QTBUG-131459
Fixes: QTBUG-134219
Change-Id: I61d560c3d984c532b9bfb5ad72dee74ea50cf510
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 816182d3cc50506529150039a05f8a94c03fe67e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>