linguist: add support for representation of labels
In ID based translation, labels are used to group different
messages. The labels are implemented in the earlier commits
in different translation tools, except linguist.
This patch adds a tab in linguist to switch between ID based and
text based translation, and grouping the messages based on the
translation type (either context in text based, or label in
ID based).

Change-Id: I3778fb017e43a73d184e92c8ac95e15a2b32aef3
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>