QDoc: Strip double quotes that wrap `\image` text
The QDoc manual contains wrong information about how to correctly use
the `\image` command, which has led to widespread use of curly braces
and double quotes in the string that's used as alternate text for the
image. This string is protected, which means HTML entity codes end up
as part of the string. The manual also suggests multi-line image text
should be wrapped in curly braces; however, these also end up as part
of the output.

This change allows QDoc to detect when a description string is wrapped
in double quotes. When detected, these double quotes are removed from
the string. Also, the change allows the `\image` command to take the
textual description as an argument in braces. The documentation is
updated accordingly, now recommending the use of braces for the textual
description. A test, introduced for the related task that made this
buggy behavior apparent, is extended to cover the various ways of using
the command (within reason).

Fixes: QTBUG-134520
Task-number: QTBUG-132432
Change-Id: Ic634c0b6b0978ef929c895e656e4eef8f4ff3de9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 5a6e9667bfefb6b1955cd8d5c95f399dc7990e59)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>