CMake: Fix location of generated .qm files
If a project calls qt_add_translations in a subdirectory, the call can
be deferred to the end of the project scope. In that case, the default
output directory for .qm files was PROJECT_BINARY_DIR. But it should be
CMAKE_CURRENT_BINARY_DIR of the subdirectory.

Fix this by correctly setting the default value of QM_OUTPUT_DIRECTORY
before forwarding arguments to the deferred call.

Pick-to: 6.9.0
Fixes: QTBUG-134693
Change-Id: I50f91fc29141730ea82f1d8968779a0934646e22
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
(cherry picked from commit 42d263544cf1da295fcd711dcda5edab8829c558)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>