designer: minimal support for labels in older versions
The support for labels in linguist and designer are provided in patches:
3c0836a55d
d6ec2c354f
a304298882
If9f010f84

Though with the introduction of labels, older versions of designer
generate an error on UI files with labels.
This patch provides minimal changes to be picked on the older LTS
versions, to avoid generating errors in those older versions.

Pick-to: 6.8
Pick-to: 6.9
Pick-to: 6.10
Change-Id: I8197a16db7cd566fe5b3a01f7858cdcc214dbb13
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>