qdoc: Allow generated lists in table of contents structure
Some documentation projects already use \generatelist or \annotatedlist
to list group member pages, and have the output act as the table of
contents (TOC) for the project.

A limitation of this approach was that if such TOC pages were added
to `navigation.toctitles` qdocconf variable, navigation links (previous/
next page) were only created for explicit link (\l) commands, but not
for the auto-generated group members.

This change adds that support. In QDocDatabase::updateNavigation(),
process also Atom types `GeneratedList` and `AnnotatedList`, fetch
the members of the associated group, and update their nav. links using
the same logic as is done for explicit links in list items.

The only practical difference is that for group members, the navigational
parent is *not* updated. This is because the navigational parent of a
group member should always be the \group page that owns it, and
overwriting that could be unexpected or confusing.

Document the new feature and add a test for it.

[ChangeLog][QDoc] Pages that contain table-of-contents \list structures
can now also use \generatelist and \annotatedlist commands for the same
purpose.

Pick-to: 6.10
Fixes: QTBUG-128268
Change-Id: I29dd71eb87d4da512ecc92a312d52a4a20e46305
Reviewed-by: David Boddie <david.boddie@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>