QDoc: Drop 32-bit support in CodeMarker
QDoc's CodeMarker class includes two functions that, when used in
tandem, serialize and deserialize a `Node *` to/from a QString. The
`nodeForString` function includes logic that obtains a different integer
depending on whether the underlying system is 32-bit or 64-bit. This
allows QDoc to run on both 32-bit and 64-bit operating systems.

QDoc's operating support follows that of Qt, and only for the three
major desktop platforms; Linux, macOS, and Windows. In practice, new
deployments or updates that adhere to Qt's OS support policies will use
64-bit environments. This means QDoc's 32-bit support should no longer
be considered a requirement. Observe the following:

- Windows: Windows 10, while technically still supported, is scheduled
  to be end-of-life in October 2025, shortly after the scheduled release
  of Qt 6.10, which is 23.09.2025. Windows 11 is 64-bit only. Thus,
  32-bit support on Windows is shrinking fast.
- macOS: Apple dropped 32-bit support entirely with macOS Catalina back
  in 2019, six years ago. All supported macOS platforms are exclusively
  64-bit.
- Linux: Most major platforms (e.g. Debian, Ubuntu, Fedora) have
  either deprecated or dropped full 32-bit support.

This change removes the architecture-specific code in
`CodeMarker::nodeForString()`. The function now returns a `Node *`
from an unsigned long long, always.

[ChangeLog][QDoc] - QDoc no longer supports 32-bit platforms.

Change-Id: I38d76bc7510fd3d299528fff9276fc426fa82107
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: David Boddie <david.boddie@qt.io>