QDoc: Quickly determine that a PageNode is used to avoid an assert
The beginSubPage() function contained an assert that the node supplied
to it was a PageNode. Since it was possible to reason that all places
in Generator call it with a PageNode or instance of a subclass, the
signatures of beginSubPage() and other functions were updated to use
PageNode instead of Node.

Change-Id: I9d2c9f8e580f9d4986714e5b4586cfd0456b86dd
Reviewed-by: Paul Wicking <paul.wicking@qt.io>