qdoc: Fix linking to output subdirectories in single-exec mode
When constructing a link to a Node that lives outside the project
currently being generated, QDoc checked whether the link target is an
"index node". If so, the href of the link was prepended with
`../<project-name>` of the target Node. This works as expected in normal
operation.

However, in single-exec mode, testing for index nodes is always false,
as nothing is loaded from index files. This resulted in incorrect
links unless all documentation projects that are part of the same
single-exec build were configured to use the same output (sub)dir.

Pick-to: 6.9 6.8
Fixes: QTBUG-136115
Change-Id: Id6624c52f5190e521cecbfc3176f8ac066590daf
Reviewed-by: Paul Wicking <paul.wicking@qt.io>