Doc: Update the \modulestate command
There are different states that a module can be in, for example, an
'active' state, a deprecated state, marked for deprecation, or a custom
state that the user can define.

Whilst the \modulestate command can be used with an argument, for
example, deprecated, QDoc will not actually deprecate the module. It
will just generate a sentence on the module page. This can cause confusion
for the user.

This patch expands on the \modulestate command and explains that it can be
used for custom module state descriptions. A note was added to state
that it should not be used to actually deprecate a module.

Fixes: QTBUG-133739
Pick-to: 6.8
Change-Id: I4d439a59d6f4fbee4506ea2a297df107853992da
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
(cherry picked from commit 78be281374fcc6959f4d386ec7b08d1cc90c0422)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>