QDoc: Unify parsing code for image commands
QDoc offers two image commands; CMD_IMAGE and CMD_INLINEIMAGE. They
differ in their purpose and thus also their behavior in that the former
always breaks out of list structures etc, while the latter doesn't -- it
'inlines' images.

However, apart from this, the code is mostly the same, with a notable
difference: The code for CMD_IMAGE has been modified to make QDoc issue
warnings for images that contain no alt text. This leads to a subtle
and unexpected difference in behavior between the two.

This patch unifies the code path for the two commands, such that they
behave identically except where they must behave differently. This
should result in more coherent behavior and hopefully reduce the chance
that the behavior of the two diverges again in the future, as
maintenance will be easier.

Fixes: QTBUG-136436
Change-Id: Ibcdbb322dc11f3c7ed4ef16216ceb3e2e022dcdc
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>