QDoc: Prevent warning on explicitly empty image alt text
QDoc allows users to embed images in documentation by way of two
commands, `\image` and `\inlineimage`. The parsing code for the two
commands was recently unified; as a result, both commands now may
trigger a warning from QDoc about missing alt text attributes for
images. However, the implementation of the warning fails to account for
alt text attributes that are missing or empty on purpose. This is often
desireable for an inline image, which may, for example, be an icon (or
other user interface element) that is described by the text that
surrounds it.

This change modifies QDoc such that it will no longer warn about a
missing alt text if the command is used in a way that is interpreted as
indicating intent. For example, the following commands will result in a
warning:

    \image foo.jpg
    \image {foo.jpg}
    \inlineimage bar.jpg "some alt text"
    \inlineimage {bar.jpg}

while these will not:

    \image foo.jpg ""
    \image {foo.jpg} {}
    \inlineimage {bar.jpg} {}

A test case for images in the end-to-end test that validates QDoc's
output is modified to account for and exercise the change in
behavior. To keep the test running without warnings, the configuration
flag that enables the issuing of the warning is left unset, but also
referenced in the test text as a heads up to the observant reader.

Task-number: QTBUG-136436
Change-Id: Id8d645c30bd8c2fc14892fd744fdb1cfb8fefea7
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>