QDoc: Reorganize includes in Utilities
The dependencies of `Utilities` is about to grow. This change
reorganizes the #include directives such that they follow the
conventional style. This is made as an atomic change, as it's not
related to the changes that are to follow, and getting this out of the
way first reduces git friction.

Change-Id: I803f63c9c25cb6dc40a4f6a3612835c28e1ae0ad
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>