QDoc: Put Node serialization functions in Utilities
QDoc's `CodeMarker` class provides a pair of static utility functions,
`nodeForString` and `stringForNode`. These are used to convert between a
`Node` pointer and its string representation (based on the pointer
address). As the functions are not directly related to code marking, the
location is sub-optimal.

This change moves these two functions into the `Utilities` namespace,
placing them in a more appropriate location for general utility
code. The interface has been modified to make use of the new `INode`
base class for `Node`. This improves code organization, separation of
concerns, and avoids making `Utilities` depend on `Node`. All call
sites have been updated to use `Utilities::nodeForString` and
`Utilities::stringForNode`.

As the functions weren't documented, this change also adds documentation
for them. Finally, as the purpose of the Utilities namespace since its
inception was to offer a more easily tested place for utility
functions, the change adds a set of simple unit tests for them both.

Change-Id: Ib90a1d97bbd45318c7142d94b591f0ccfd104221
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>