QDoc: Clean up some includes
QDoc's `Node` class is a bit of a "God class" and its implementation
details leak across QDoc's code base. One example of this is how the
`Atom` class contains an include directive for the `node.h` header,
despite the fact that Atom doesn't operate on (and thus doesn't need to
know the implementation of) Node.

By removing the include directive, a range of transient includes becomes
apparent; includes that are no longer satisfied when Atom doesn't
already carry Node around. This change seeks to remedy that situation,
by removing the include directive in Atom, and adding forward
declarations where possible, actual includes where necessary. The
purpose of this change is to decouple Node from other parts of QDoc in
the quest for a more testable code base.

Task-number: QTBUG-135943
Change-Id: I8a3a0186579adf15073f760e70f575691a2ab2f4
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>