QDoc: Viva the Generator::addNodeLink functions!
QDoc's generators constructs link formatting atoms in multiple
locations. This results in repetitive Atom sequences scattered through
the implementation.

This change introduces two free functions to QDoc's `Generator`
class, one of which is a convenience overload for the other. They serve
as a way to abstract the repetitive Atom sequences. The change replaces
the candidate sequences with calls to the new function(s). This is
intended to improve future maintainability by placing part of the link
creation logic in a central location. The intention is also to improve
code readability by stowing away link creation details.

In one call site in `HTMLGenerator`, there's a minor code style
improvement by moving to an if-init statement for one of the impacted
branches.

Task-number: QTBUG-92109
Change-Id: I453c36707f884a19fd035256c2bd84cb3d82d590
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>