QDoc: Wrap GenusTypes in QT_BEGIN|END_NAMESPACE
QDoc's enums `Genus` and `NodeType` are declared in the global
namespace. This can potentially lead to name clashes when QDoc is built
as part of a namespaced Qt build, if other code defines types with the
same names. This can also cause issues for other parts of QDoc that
e.g. forward declare the types.

This change wraps the content of `genustypes.h` within
`QT_BEGIN_NAMESPACE` and `QT_END_NAMESPACE` macros. This ensures that
these types are properly scoped within the Qt namespace, mitigating the
risk of naming conflicts.

Task-number: QTBUG-135943
Change-Id: Ic6006f6f4d0fab9a3e798ea0f283c945f1803082
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>