QDoc: Rename a local variable
A recent change introduced the `addNodeLink` functions to QDoc's
Generator class. That change extracted a common pattern into a set of
convenience functions. In the updates to the call sites, a subtle bug
snuck in, which is the use of the lowercase `ns` symbol where the
uppercase `NS` should have been used. One of QDoc's end-to-end tests
caught the regression before it entered the code base, but this exposed
the issue of similar (indeed, one could say overlapping) two-letter
symbol names within the same chunk of code.

This change renames one of these symbols in an attempt to make the code
more expressive. Hopefully, this will also aid future developers in
making the same refactoring mistake in this part of QDoc's code.

Task-number: QTBUG-92109
Change-Id: I73b01ed1f5da5a39b4904c53267e190cf90e4080
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>