QDoc: Add regression test for macro expansion in links
QDoc's `\l` command accepts QDoc macro as the link target, and will
expand the macro in place before linking to the target. As multi-word
targets must be wrapped in curly braces, macros that expand to multiple
words (i.e. a string that contains spaces) must also be wrapped in curly
braces.

This commit adds a test case to QDoc's end-to-end test that verifies
QDoc's ability to link to macro expansions using the `\l` command. The
test serves as documentation of this functionality and prevents future
regressions when modifying QDoc's link processing.

Pick-to: 6.9 6.8
Change-Id: I29efc9b419569946e5dcf7d64dfea3319944092a
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>