QDoc: Move overload command handling in DocParser
When QDoc parses a documentation comment, it performs a range of
operations. This includes registering and processing certain
"meta-commands", including `CMD_OVERLOAD`. The implementation is buggy:

  - It causes QDoc to generate overload text for functions even when
    they are not actually overloaded.
  - It generates the overload text in place; an accidentally placed
    `\overload` inside of a table, for example, will generate the text
    inside the table. This behavior is inconsistent and surprising.

This change is a pure refactoring change that moves the code responsible
for processing the `\overload` command into a new helper function. The
parser is updated to delegate to the new function. This is done as a
separate step for two reasons:

  - Delegating to a helper function delegates the responsibility away
    from `DocParser::parse()`. This is leads to a cleaner architecture
    and improves maintainability.
  - The refactoring helps reduce the surface in the change that
    addresses the faulty behavior. This (hopefully) reduces the
    complexity of the change, making it easier to review the bugfix
    change for correctness.

This change also adds documentation for the new helper function.

Pick-to: 6.8 6.9.1
Task-number: QTBUG-136532
Change-Id: I8cf0a2e7034868cfdef85034390ced036e821413
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit f35370ea4cfc17f59f12fecb51e0b5d8997813c3)