QDoc: Differentiate overload notes from admonitions
A recent change to QDoc improved QDoc's handling of the `\overload`
command. However, as an unintentional consequence, overload notes are
now prefixed with "Note:", and get the same visual treatment as other
admonitions. This makes them appear as important or even critical
information.

What allowed this to happen and pass review, is identified as the API
in `Generator::generateAddendum()`, which takes a boolean value as the
fourth parameter. This boolean value controls the choice of CSS class
and generation of the "Note:" prefix. An overload that only takes three
parameters and unconditionally passes `true` for the boolean allowed the
issue to sneak in.

This change introduces an `AdmonitionPrefix` enum to replace the boolean
parameter that controlled prefix display. The overload is modified to
pass `AdmonitionPrefix::Note`, such that the original behavior is kept
for calls to the function version that takes three arguments. Other call
sites have been updated accordingly. As a result, overload notes now
appear without the "Note:" prefix while maintaining their informational
content. Test data is updated to reflect the new behavior.

Pick-to: 6.8 6.9.1
Task-number: QTBUG-136532
Change-Id: I39b923d6a4079079d5071136f13273032f279029
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 7031d090fe0e2620de1e984dad6cba7682a477d3)