qdoc: Introduce \qmlenum command
Introduce and document a new QDoc topic command, \qmlenum, for
documenting enumerations in QML.

Add CppCodeParser::processQmlEnumTopic() to do the work; parse
the command argument with QmlPropertyArguments, and on success,
construct a new QmlEnumNode instance.

Add all documented enumerators to the node; this ensures linking
by type + enumerator name works the same way as it does for C++
\enum.

Note that QDoc's QML parser does not support generating
enumerations as there is no declaration an enumeration could be
tied to in a .qml file. Therefore, the documentation must live in
a .qdoc or a .cpp source file.

[ChangeLog][QDoc] QDoc now supports documenting enumerations in QML
using the \qmlenum command.

Task-number: QTBUG-134806
Change-Id: I0677e92933c85beaae1f046d597d50079ad9de0b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Safiyyah Moosa <safiyyah.moosa@qt.io>