qdoc: Add QmlEnumNode instances to index
Write QmlEnumNode instances to .index files, so they can be linked to
from external documentation projects.

As QmlEnumNode derives from EnumNode, mostly reuse the same code path.
The only differences are the element name ("qmlenum"), and the fact
that enumerator values are not stored for QML enumerations; these values
are not valid (nor are they displayed in the documentation).

Task-number: QTBUG-134806
Change-Id: I7987fb030acd8bb06f0706d92c9aab6d96ae50ac
Reviewed-by: Paul Wicking <paul.wicking@qt.io>